@@ -4,7 +4,7 @@
 @{ 
   this.ViewBag.Title = "View from Extension A";
 }
-<h1>View from Extension A (Added as Resource)</h1>
+<h1>View from Extension A (<span class="suspended">Added as Resource</span>* Precompiled)</h1>
 <p class="from-resources">This line is red because of the rule in CSS-file added as resource (from Extension A).</p>
 <p>Installed extensions:</p>
 <ul>
@@ -14,3 +14,4 @@
 }
 </ul>
 <p><a href="/extension-b">Go to Extension B</a></p>
+<p>* Razor runtime compilation is turned off by default starting from ASP.NET Core 3.0 and not supported by the ExtCore framework so far (5.0.0-alpha1).</p>
